gk-deploy: fix error in check_pods() when second argument is a string
The second argument of check_pods() can either be a timeout integer
value or a string to check for the status field. In the case of
a status string, it was mistakenly used as an override for the
timeout, leading to always error with timeout.

This patch lets check_pods inspect the second argument and only
use it as a timout override if it is numeric.

Signed-off-by: Michael Adam <obnox@redhat.com>