when using ssh executor provide /etc/fstab as ENV
If ssh executor is used, it also generally means that Gluster is not
containerized. In such instances, we don't have the startup script in
Gluster container which would look at /var/lib/heketi/fstab for bricks.
Hence, this patch changes the HEKETI_FSTAB ENV to /etc/fstab when
sshexec is chosen.

For now, this is sufficient. When and if we want to support ssh access
to Gluster containers, we will have to provide a flag --fstab to let
users decide where the file should reside. Not now.

Signed-off-by: Raghavendra Talur <rtalur@redhat.com>