vagrant: add note to the main readme to indicate caching is possible
The quickstart guide is nice and straightforward but I didn't intially
realize caching was possible. This change adds a note in the quickstart
to indicate to readers that caching is possible and that one should
refer to the vagrant directory readme for that information.

I was a tad bit iffy on where to put the note so I'm willing to move
it to later in the quickstart if we think it interrupts the flow
of the text.

Signed-off-by: John Mulligan <jmulligan@redhat.com>