vagrant: ./up.sh should use --no-parallel to prevent download issues
When I run ./up.sh in a clean environment, at least one of the VMs does
not come up. The following error is then displayed:

    An error occurred while executing multiple actions in parallel.
    Any errors that occurred are shown below.

    An error occurred while executing the action on the 'node0'
    machine. Please handle this error then try again:

    The box 'centos/7' could not be found or
    could not be accessed in the remote catalog. If this is a private
    box on HashiCorp's Atlas, please verify you're logged in via
    `vagrant login`. Also, please double-check the name. The expanded
    URL and error message are shown below:

    URL: https://atlas.hashicorp.com/centos/7
    Error: Could not resolve host: atlas.hashicorp.com

In order to keep the `vagrant up` running in parallel, a check for the
availability of the box is added. In case the box is not available, it
will be downloaded before `vagrant up` is executed.

Unfortunately Vagrant will ask for the provider that will be used. It
seems not possible to pass a `--provider=libvirt` on the commandline, or
set the `VAGRANT_DEFAULT_PROVIDER=libvirt` environment variable. Neither
prevents the input from being asked.

Suggested-by: Jose A. Rivera <jarrpa@redhat.com>
Signed-off-by: Niels de Vos <ndevos@redhat.com>