wait for kubernetes before trying to deploy weave
This fixes an issue I was hitting with mutltiple runs. Usually the
first run of the playbooks would work fine but if a different issue
was hit and I neded to re-run them I kept hitting errors on
the weave step related to connection failures. This change adds a
wait for the port to be available before running that troublesome
step.

Signed-off-by: John Mulligan <jmulligan@redhat.com>