vagrant: use a matching version of Weave Net
While deploying the Vagrant environment, Kubernetes 1.9.2 is currently
used. This version does not seem to be compatible with the 1.6 version
of Weave Net. While the weave- pods are starting, the following errors
are reported:

  dockerd-current[6461]: ERROR: logging before flag.Parse: E0119 15:50:15.165210    9894 reflector.go:205] github.com/weaveworks/weave/prog/weave-npc/main.go:230: Failed to list *v1.NetworkPolicy: networkpolicies.networking.k8s.io is forbidden: User "system:serviceaccount:kube-system:weave-net" cannot list networkpolicies.networking.k8s.io at the cluster scope

By using the recommended way to install the add-on, a matching version
will automatically be picked.

URL: https://www.weave.works/docs/net/latest/kubernetes/kube-addon/
Suggested-by: John Mulligan <jmulligan@redhat.com>
Signed-off-by: Niels de Vos <ndevos@redhat.com>