gk-deploy: check if heketi in the pod can support the --single-node option
Currently there is no Heketi release available that supports the
--durability option for the 'heketi-cli setup-openshift-heketi-storage'
command. In order to give the users a meaningful error message, add a
check to see if the --durability=none option is accepted. This is done
in combination with the --help option, which will just exit successfully
in case the options are valid, or exit with an error otherwise.

Signed-off-by: Niels de Vos <ndevos@redhat.com>