deploy: enable checking for running gluster-blockd service
The probe script does not check for gluster-blockd by default. Only when
setting the GLUSTER_BLOCKD_STATUS_PROBE_ENABLE environment variable to
"1" the check is activated.

New deployments with gk-deploy should have everything that makes
gluster-blockd available. In environments where this is not the case,
the GLUSTER_BLOCKD_STATUS_PROBE_ENABLE can be set to "0" to disable the
probe for gluster-blockd.

Signed-off-by: Niels de Vos <ndevos@redhat.com>