deploy: bind-mount /dev/disk into the glusterfs-server container
Commit 2f1114a removed the bind-mount for /dev as this will be setup by
the container runtime automatically. However it seems that /dev/disk is
missing, prevending users to supply /dev/disk/by-id/* device names to
Heketi.

This change adds /dev/disk as an additional bind-mount. CRI-O does
prevent /dev to be bind-mounted, but subdirectories of /dev do not seem
to be an issue.

Reported-by: Gal Ben Haim <gbenhaim@redhat.com>
Signed-off-by: Niels de Vos <ndevos@redhat.com>