deploy: provide an alternative volume mount for /dev
The gluster-setup.sh script in the glusterfs-server container now chceks
for the HOST_DEV_DIR environment variable and will 'mount --rbind' the
alternative path to /dev from the host to /dev in the container. This is
needed to allow access to *all* devices on the host, including ones that
get dynamically allocated or hot-plugged.

See-also: gluster/gluster-containers#115
Signed-off-by: Niels de Vos <ndevos@redhat.com>