Correcting the modules path in order to support other distros
This patch fixes #539. Not all supported distros have /lib as a symlink of
/usr/lib. For example, Ubuntu 18.04 doesn't. This causes an issue with
tcmu-runner; which requires target_core_user and other modules to be able to
build.

In order to fix this, we just need to mount `/lib/modules` instead so all of
the supported distributions are capable of the build.