deploy: add HEKETI_DEBUG_UMOUNT_FAILURES environment variable for extra debugging
Heketi offers an additional way to debug troubles in case unmounting a
brick of a Gluster volume fails. This is a problem that is not observed
often or consistently. However debugging the issue is difficult as open
file descriptors can get closed at a later point, showing no trace of
the origin of the failure. By setting the HEKETI_DEBUG_UMOUNT_FAILURES
variable to "true", additional information will be available in the logs
of the Heketi pod.

See-also: heketi/heketi#1483
Signed-off-by: Niels de Vos <ndevos@redhat.com>